0069-SQRTX Stats: Time: 0 ms (100%), Space: 8.7 MB (14.71%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
     int mySqrt(int x) {
-        return static_cast<int>(sqrt(x));
+        return sqrt(x);
     }
 };
